+++ lib/transpec/configuration.rb
@@ -5,15 +5,15 @@ module Transpec
     NEGATIVE_FORMS_OF_TO = ['not_to', 'to_not'].freeze
 
     PREDICATES = [
-      [:convert_to_expect_to_matcher, true],
-      [:convert_to_expect_to_receive, true],
-      [:convert_to_allow_to_receive,  true],
-      [:convert_have_items,           true],
-      [:convert_deprecated_method,    true],
-      [:parenthesize_matcher_arg,     true],
-      [:forced,                       false],
-      [:skip_dynamic_analysis,        false],
-      [:generate_commit_message,      false]
+      [:convert_should,            true],
+      [:convert_should_receive,    true],
+      [:convert_stub,              true],
+      [:convert_have_items,        true],
+      [:convert_deprecated_method, true],
+      [:parenthesize_matcher_arg,  true],
+      [:forced,                    false],
+      [:skip_dynamic_analysis,     false],
+      [:generate_commit_message,   false]
     ].freeze
 
     PREDICATES.each do |predicate, _|